popolamento databhase dei loghi

+++ PROGETTOFINALE/n3/CapstonebackEnd/src/main/java/nextDevs/CapstonebackEnd/model/Logo.java
@@ -2,15 +2,18 @@ package nextDevs.CapstonebackEnd.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+
 import jakarta.persistence.Id;
 import lombok.Data;
+
+import java.io.Serial;
 
 @Data
 @Entity
 public class Logo {
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
     private String symbol;
     private String url;
 }
